LayoutTests/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-pseudo-expected.txt

Item

PASS Resolution of width is correct when pseudo-element argument is ignored (due to no colon)
PASS Resolution of width is correct when pseudo-element argument is invalid (due to a trailing token)
PASS Resolution of width is correct for ::before and ::after pseudo-elements (single-colon)
PASS Resolution of width is correct for ::before and ::after pseudo-elements (double-colon)
PASS Pseudo-elements can use the full range of CSS syntax
PASS Resolution of width is correct for ::before and ::after pseudo-elements of display: contents elements
PASS Resolution of nonexistent pseudo-element styles
PASS Resolution of pseudo-element styles in display: none elements
PASS Item-based blockification of pseudo-elements
FAIL Item-based blockification of nonexistent pseudo-elements assert_equals: Pseudo-styles of display: flex elements should get blockified expected "block" but got "inline"
PASS display: contents on pseudo-elements
PASS Dynamically change to display: contents on pseudo-elements
PASS Unknown pseudo-elements
PASS CSSStyleDeclaration is immutable
PASS Unknown pseudo-element with a known identifier: backdrop
FAIL Unknown pseudo-element with a known identifier: file-selector-button assert_equals: Should return the ::file-selector-button style expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
PASS Unknown pseudo-element with a known identifier: grammar-error
FAIL Unknown pseudo-element with a known identifier: highlight(name) assert_equals: Should return the ::highlight(name) style expected "rgb(0, 128, 0)" but got ""
PASS Unknown pseudo-element with a known identifier: marker
FAIL Unknown pseudo-element with a known identifier: placeholder assert_equals: Should return the ::placeholder style expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
PASS Unknown pseudo-element with a known identifier: spelling-error
PASS Unknown pseudo-element with a known identifier: view-transition
FAIL Unknown pseudo-element with a known identifier: view-transition-image-pair(name) assert_equals: Should return the ::view-transition-image-pair(name) style expected "rgb(0, 128, 0)" but got ""
FAIL Unknown pseudo-element with a known identifier: view-transition-group(name) assert_equals: Should return the ::view-transition-group(name) style expected "rgb(0, 128, 0)" but got ""
FAIL Unknown pseudo-element with a known identifier: view-transition-old(name) assert_equals: Should return the ::view-transition-old(name) style expected "rgb(0, 128, 0)" but got ""
FAIL Unknown pseudo-element with a known identifier: view-transition-new(name) assert_equals: Should return the ::view-transition-new(name) style expected "rgb(0, 128, 0)" but got ""
